requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
aiohttp== 3.8.5
SQLAlchemy==2.0.20
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
asyncpg==0.28.0
psycopg2-binary==2.9.7
fastapi-cache2==0.2.1
redis==4.6.0
jinja2==3.1.2
gunicorn